@@ -1,5 +1,5 @@
-class AddIndexToUsersEmail < ActiveRecord::Migration
+class AddIndexToUsersEmail < ActiveRecord::Migration[4.2]
   def change
-  	add_index :users, :email, unique: true
+    add_index :users, :email, unique: true
   end
 end
