@@ -3,7 +3,7 @@
 # bundle exec rake db:migrate
 # bundle exec rake test:prepare
 
-class AddAdminToUsers < ActiveRecord::Migration
+class AddAdminToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :admin, :boolean, default: false
   end
