@@ -1,40 +1,38 @@
 require 'spec_helper'
+require 'rails_helper'
 
-describe RelationshipsController do
-
-  let(:user) { FactoryGirl.create(:user) }
-  let(:other_user) { FactoryGirl.create(:user) }
+describe RelationshipsController, type: :controller do
+  let(:user) { FactoryBot.create(:user) }
+  let(:other_user) { FactoryBot.create(:user) }
 
   before { sign_in user, no_capybara: true }
 
-  describe "creating a relationship with Ajax" do
-
-    it "should increment the Relationship count" do
+  describe 'creating a relationship with Ajax' do
+    it 'should increment the Relationship count' do
       expect do
-        xhr :post, :create, relationship: { followed_id: other_user.id }
+        post :create, params: { relationship: { followed_id: other_user.id } }, xhr: true
       end.to change(Relationship, :count).by(1)
     end
 
-    it "should respond with success" do
-      xhr :post, :create, relationship: { followed_id: other_user.id }
-      expect(response).to be_success
+    it 'should respond with success' do
+      post :create, params: { relationship: { followed_id: other_user.id } }, xhr: true
+      expect(response).to be_successful
     end
   end
 
-  describe "destroying a relationship with Ajax" do
-
+  describe 'destroying a relationship with Ajax' do
     before { user.follow!(other_user) }
     let(:relationship) { user.relationships.find_by(followed_id: other_user) }
 
-    it "should decrement the Relationship count" do
+    it 'should decrement the Relationship count' do
       expect do
-        xhr :delete, :destroy, id: relationship.id
+        post :destroy, params: { id: relationship.id }, xhr: true
       end.to change(Relationship, :count).by(-1)
     end
 
-    it "should respond with success" do
-      xhr :delete, :destroy, id: relationship.id
-      expect(response).to be_success
+    it 'should respond with success' do
+      post :destroy, params: { id: relationship.id }, xhr: true
+      expect(response).to be_successful
     end
   end
-end 
+end
