@@ -1,195 +1,190 @@
 require 'spec_helper'
 
-describe "Authentication" do
+describe 'Authentication' do
+  subject { page }
 
-	subject { page }
+  describe 'signin page' do
+    before { visit signin_path }
 
-	describe "signin page" do
-		before { visit signin_path }
+    it { should have_content('Sign in') }
+    it { should have_title('Sign in') }
+  end
 
-		it { should have_content('Sign in') }
-		it { should have_title('Sign in') }
-	end
+  describe 'signin' do
+    before { visit signin_path }
 
-	describe "signin" do
-		before { visit signin_path }
+    describe 'with invalid information' do
+      before { click_button 'Sign in' }
 
-		describe "with invalid information" do
-			before { click_button "Sign in" }
+      it { should have_title('Sign in') }
+      it { should have_selector('div.alert.alert-error', text: 'Invalid') }
+    end
 
-			it { should have_title('Sign in') }
-			it { should have_selector('div.alert.alert-error', text: 'Invalid') }
-		end
+    describe 'after visiting another page' do
+      before { click_link 'Home' }
+      it { should_not have_selector('div.alert.alert-error') }
+    end
 
-		describe "after visiting another page" do
-			before { click_link "Home" }
-			it { should_not have_selector('div.alert.alert-error') }
-		end
+    describe 'with valid information' do
+      let(:user) { FactoryBot.create(:user) }
+      # before do
+      # 	fill_in "Email",    with: user.email.upcase
+      # 	fill_in "Password", with: user.password
+      # 	click_button "Sign in"
+      # end
+      # Change to the spec/support/utilities.rb
+      before { sign_in user }
 
-		describe "with valid information" do
-			let(:user) { FactoryGirl.create(:user) }
-			# before do
-			# 	fill_in "Email",    with: user.email.upcase
-			# 	fill_in "Password", with: user.password
-			# 	click_button "Sign in"
-			# end
-			# Change to the spec/support/utilities.rb
-			before { sign_in user }
+      it { should have_title(user.name) }
+      it { should have_link('Users',       href: users_path) }
+      it { should have_link('Profile',     href: user_path(user)) }
+      it { should have_link('Settings',    href: edit_user_path(user)) }
+      it { should have_link('Sign out',    href: signout_path) }
+      it { should_not have_link('Sign in', href: signin_path) }
 
-			it { should have_title(user.name) }
-			it { should have_link('Users',       href: users_path) }
-			it { should have_link('Profile',     href: user_path(user)) }
-			it { should have_link('Settings',    href: edit_user_path(user)) }
-			it { should have_link('Sign out',    href: signout_path) }
-			it { should_not have_link('Sign in', href: signin_path) }
+      describe 'followed by signout' do
+        before { click_link 'Sign out' }
+        it { should have_link('Sign in') }
+      end
+    end
+  end
 
-			describe "followed by signout" do
-				before { click_link "Sign out" }
-				it { should have_link('Sign in') }
-			end
-		end
-	end
+  describe 'authorization' do
+    describe 'for non-signed-in users' do
+      let(:user) { FactoryBot.create(:user) }
 
-	describe "authorization" do
+      describe 'in the Users controller' do
+        describe 'visiting the edit page' do
+          before { visit edit_user_path(user) }
+          it { should have_title('Sign in') }
+        end
 
-		describe "for non-signed-in users" do
-			let(:user) { FactoryGirl.create(:user) }
+        # apart from Capybara’s visit method, to access a controller action: by issuing the appropriate HTTP request directly,
+        # in this case using the patch method to issue a PATCH request:
+        describe 'submitting to the update action' do
+          before { patch user_path(user) }
+          specify { expect(response).to redirect_to(signin_path) }
+        end
 
-			describe "in the Users controller" do
+        describe 'visiting the user index' do
+          before { visit users_path }
+          it { should have_title('Sign in') }
+        end
 
-				describe "visiting the edit page" do
-					before { visit edit_user_path(user) }
-					it { should have_title('Sign in') }
-				end
+        describe 'visiting the following page' do
+          before { visit following_user_path(user) }
+          it { should have_title('Sign in') }
+        end
 
-				# apart from Capybara’s visit method, to access a controller action: by issuing the appropriate HTTP request directly, 
-				# in this case using the patch method to issue a PATCH request:
-				describe "submitting to the update action" do
-					before { patch user_path(user) }
-					specify { expect(response).to redirect_to(signin_path) }
-				end
+        describe 'visiting the followers page' do
+          before { visit followers_user_path(user) }
+          it { should have_title('Sign in') }
+        end
+      end
 
-				describe "visiting the user index" do
-					before { visit users_path }
-					it { should have_title('Sign in') }
-				end
+      describe 'when attempting to visit a protected page' do
+        before do
+          visit edit_user_path(user)
+          fill_in 'Email',    with: user.email
+          fill_in 'Password', with: user.password
+          click_button 'Sign in'
+        end
 
-				describe "visiting the following page" do
-					before { visit following_user_path(user) }
-					it { should have_title('Sign in') }
-				end
+        describe 'after signing in' do
+          # Should redirect back to Edit page after re-signin
+          it 'should render the desired protected page' do
+            expect(page).to have_title('Edit user')
+          end
+        end
+      end
 
-				describe "visiting the followers page" do
-					before { visit followers_user_path(user) }
-					it { should have_title('Sign in') }
-				end
-			end
+      # describe "when attempting to visit a protected page" do
+      # 	before do
+      # 		visit edit_user_path(user)
+      # 		fill_in "Email",    with: user.email
+      # 		fill_in "Password", with: user.password
+      # 		click_button "Sign in"
+      # 	end
 
-			describe "when attempting to visit a protected page" do
-				before do
-					visit edit_user_path(user)
-					fill_in "Email",    with: user.email
-					fill_in "Password", with: user.password
-					click_button "Sign in"
-				end
+      # 	describe "after signing in" do
 
-				describe "after signing in" do
-					# Should redirect back to Edit page after re-signin
-					it "should render the desired protected page" do
-						expect(page).to have_title('Edit user')
-					end
-				end
-			end
+      # 		it "should render the desired protected page" do
+      # 			expect(page).to have_title('Edit user')
+      # 		end
 
-			# describe "when attempting to visit a protected page" do
-			# 	before do
-			# 		visit edit_user_path(user)
-			# 		fill_in "Email",    with: user.email
-			# 		fill_in "Password", with: user.password
-			# 		click_button "Sign in"
-			# 	end
+      # 		describe "when signing in again" do
+      # 			before do
+      # 				delete signout_path
+      # 				visit signin_path
+      # 				fill_in "Email",    with: user.email
+      # 				fill_in "Password", with: user.password
+      # 				click_button "Sign in"
+      # 			end
 
-			# 	describe "after signing in" do
+      # 			it "should render the default (profile) page" do
+      # 				expect(page).to have_title(user.name)
+      # 			end
+      # 		end
+      # 	end
+      # end
 
-			# 		it "should render the desired protected page" do
-			# 			expect(page).to have_title('Edit user')
-			# 		end
+      describe 'in the Microposts controller' do
+        describe 'submitting to the create action' do
+          before { post microposts_path }
+          specify { expect(response).to redirect_to(signin_path) }
+        end
 
-			# 		describe "when signing in again" do
-			# 			before do
-			# 				delete signout_path
-			# 				visit signin_path
-			# 				fill_in "Email",    with: user.email
-			# 				fill_in "Password", with: user.password
-			# 				click_button "Sign in"
-			# 			end
+        describe 'submitting to the destroy action' do
+          before { delete micropost_path(FactoryBot.create(:micropost)) }
+          specify { expect(response).to redirect_to(signin_path) }
+        end
+      end
+    end
 
-			# 			it "should render the default (profile) page" do
-			# 				expect(page).to have_title(user.name)
-			# 			end
-			# 		end
-			# 	end
-			# end
+    describe 'as wrong user', type: :request do
+      let(:user) { FactoryBot.create(:user) }
+      # This creates a user with a different email address from the default. The tests specify that the original user should not
+      # have access to the wrong user’s edit or update actions.
+      let(:wrong_user) { FactoryBot.create(:user, email: 'wrong@example.com') }
+      before { sign_in user, no_capybara: true }
 
-			describe "in the Microposts controller" do
+      describe 'submitting a GET request to the Users#edit action' do
+        before { get edit_user_path(wrong_user) }
+        specify { expect(response.body).not_to match(full_title('Edit user')) }
+        specify { expect(response).to redirect_to(root_url) }
+      end
 
-				describe "submitting to the create action" do
-					before { post microposts_path }
-					specify { expect(response).to redirect_to(signin_path) }
-				end
+      describe 'submitting a PATCH request to the Users#update action' do
+        before { patch user_path(wrong_user) }
+        specify { expect(response).to redirect_to(root_url) }
+      end
+    end
 
-				describe "submitting to the destroy action" do
-					before { delete micropost_path(FactoryGirl.create(:micropost)) }
-					specify { expect(response).to redirect_to(signin_path) }
-				end
-			end
-		end
+    describe 'as non-admin user', type: :request do
+      let(:user) { FactoryBot.create(:user) }
+      let(:non_admin) { FactoryBot.create(:user) }
 
-		describe "as wrong user" do
-			let(:user) { FactoryGirl.create(:user) }
-			# This creates a user with a different email address from the default. The tests specify that the original user should not 
-			# have access to the wrong user’s edit or update actions.
-			let(:wrong_user) { FactoryGirl.create(:user, email: "wrong@example.com") }
-			before { sign_in user, no_capybara: true }
+      before { sign_in non_admin, no_capybara: true }
 
-			describe "submitting a GET request to the Users#edit action" do
-				before { get edit_user_path(wrong_user) }
-				specify { expect(response.body).not_to match(full_title('Edit user')) }
-				specify { expect(response).to redirect_to(root_url) }
-			end
+      describe 'submitting a DELETE request to the Users#destroy action' do
+        before { delete user_path(user) }
+        specify { expect(response).to redirect_to(root_url) }
+      end
+    end
 
-			describe "submitting a PATCH request to the Users#update action" do
-				before { patch user_path(wrong_user) }
-				specify { expect(response).to redirect_to(root_url) }
-			end
-		end
+    describe 'in the Relationships controller' do
+      describe 'submitting to the create action' do
+        before { post relationships_path }
+        specify { expect(response).to redirect_to(signin_path) }
+      end
 
-
-		describe "as non-admin user" do
-			let(:user) { FactoryGirl.create(:user) }
-			let(:non_admin) { FactoryGirl.create(:user) }
-
-			before { sign_in non_admin, no_capybara: true }
-
-			describe "submitting a DELETE request to the Users#destroy action" do
-				before { delete user_path(user) }
-				specify { expect(response).to redirect_to(root_url) }
-			end
-		end
-
-		describe "in the Relationships controller" do
-			describe "submitting to the create action" do
-				before { post relationships_path }
-				specify { expect(response).to redirect_to(signin_path) }
-			end
-
-			describe "submitting to the destroy action" do
-				# Note that, in order to avoid the overhead of creating a virtually useless Relationship object, 
-				# the delete test hard-codes the id 1 in the named route
-				# This works because the user should be redirected before the application ever tries to retrieve the relationship with this id.
-				before { delete relationship_path(1) }
-				specify { expect(response).to redirect_to(signin_path) }
-			end
-		end
-	end
+      describe 'submitting to the destroy action' do
+        # Note that, in order to avoid the overhead of creating a virtually useless Relationship object,
+        # the delete test hard-codes the id 1 in the named route
+        # This works because the user should be redirected before the application ever tries to retrieve the relationship with this id.
+        before { delete relationship_path(1) }
+        specify { expect(response).to redirect_to(signin_path) }
+      end
+    end
+  end
 end
