@@ -1,45 +1,43 @@
 # rails generate integration_test micropost_pages
 require 'spec_helper'
 
-describe "MicropostPages" do
-	subject { page }
+describe 'MicropostPages' do
+  subject { page }
 
-	let(:user) { FactoryGirl.create(:user) }
-	before { sign_in user }
+  let(:user) { FactoryBot.create(:user) }
+  before { sign_in user }
 
-	describe "micropost creation" do
-		before { visit root_path }
+  describe 'micropost creation' do
+    before { visit root_path }
 
-		describe "with invalid information" do
+    describe 'with invalid information' do
+      it 'should not create a micropost' do
+        expect { click_button 'Post' }.not_to change(Micropost, :count)
+      end
 
-			it "should not create a micropost" do
-				expect { click_button "Post" }.not_to change(Micropost, :count)
-			end
+      describe 'error messages' do
+        before { click_button 'Post' }
+        it { should have_content('error') }
+      end
+    end
 
-			describe "error messages" do
-				before { click_button "Post" }
-				it { should have_content('error') }
-			end
-		end
+    describe 'with valid information' do
+      before { fill_in 'micropost_content', with: 'Lorem ipsum' }
+      it 'should create a micropost' do
+        expect { click_button 'Post' }.to change(Micropost, :count).by(1)
+      end
+    end
+  end
 
-		describe "with valid information" do
+  describe 'micropost destruction' do
+    before { FactoryBot.create(:micropost, user: user) }
 
-			before { fill_in 'micropost_content', with: "Lorem ipsum" }
-			it "should create a micropost" do
-				expect { click_button "Post" }.to change(Micropost, :count).by(1)
-			end
-		end
-	end
+    describe 'as correct user' do
+      before { visit root_path }
 
-	describe "micropost destruction" do
-		before { FactoryGirl.create(:micropost, user: user) }
-
-		describe "as correct user" do
-			before { visit root_path }
-
-			it "should delete a micropost" do
-				expect { click_link "delete" }.to change(Micropost, :count).by(-1)
-			end
-		end
-	end
+      it 'should delete a micropost' do
+        expect { click_link 'delete' }.to change(Micropost, :count).by(-1)
+      end
+    end
+  end
 end
