@@ -1,268 +1,261 @@
-# By issuing a PATCH request directly to the update method as shown in Listing 9.48, 
-	# verify that the admin attribute isn’t editable through the web. 
-	# Be sure to get first to Red, and then to Green. 
-	# (Hint: Your first step should be to add admin to the list of permitted parameters 
-	# in user_params.)
+# By issuing a PATCH request directly to the update method as shown in Listing 9.48,
+# verify that the admin attribute isn’t editable through the web.
+# Be sure to get first to Red, and then to Green.
+# (Hint: Your first step should be to add admin to the list of permitted parameters
+# in user_params.)
 # describe "forbidden attributes" do
 
 require 'spec_helper'
 
-describe "User pages" do
+describe 'User pages' do
+  subject { page }
 
-	subject { page }
+  describe 'profile page' do
+    let(:user) { FactoryBot.create(:user) }
+    let!(:m1) { FactoryBot.create(:micropost, user: user, content: 'Foo') }
+    let!(:m2) { FactoryBot.create(:micropost, user: user, content: 'Bar') }
 
-	describe "profile page" do
-		let(:user) { FactoryGirl.create(:user) }
-		let!(:m1) { FactoryGirl.create(:micropost, user: user, content: "Foo") }
-		let!(:m2) { FactoryGirl.create(:micropost, user: user, content: "Bar") }
+    before { visit user_path(user) }
 
-		before { visit user_path(user) }
+    it { should have_content(user.name) }
+    it { should have_title(user.name) }
 
-		it { should have_content(user.name) }
-		it { should have_title(user.name) }
+    describe 'microposts' do
+      it { should have_content(m1.content) }
+      it { should have_content(m2.content) }
+      it { should have_content(user.microposts.count) }
+    end
 
-		describe "microposts" do
-			it { should have_content(m1.content) }
-			it { should have_content(m2.content) }
-			it { should have_content(user.microposts.count) }
-		end
+    describe 'follow/unfollow buttons' do
+      let(:other_user) { FactoryBot.create(:user) }
+      before { sign_in user }
 
-		describe "follow/unfollow buttons" do
-			let(:other_user) { FactoryGirl.create(:user) }
-			before { sign_in user }
+      describe 'following a user' do
+        before { visit user_path(other_user) }
 
-			describe "following a user" do
-				before { visit user_path(other_user) }
+        it 'should increment the followed user count' do
+          expect do
+            click_button 'Follow'
+          end.to change(user.followed_users, :count).by(1)
+        end
 
-				it "should increment the followed user count" do
-					expect do
-						click_button "Follow"
-					end.to change(user.followed_users, :count).by(1)
-				end
+        it "should increment the other user's followers count" do
+          expect do
+            click_button 'Follow'
+          end.to change(other_user.followers, :count).by(1)
+        end
 
-				it "should increment the other user's followers count" do
-					expect do
-						click_button "Follow"
-					end.to change(other_user.followers, :count).by(1)
-				end
+        describe 'toggling the button' do
+          before { click_button 'Follow' }
+          it { should have_xpath("//input[@value='Unfollow']") }
+        end
+      end
 
-				describe "toggling the button" do
-					before { click_button "Follow" }
-					it { should have_xpath("//input[@value='Unfollow']") }
-				end
-			end
+      describe 'unfollowing a user' do
+        before do
+          user.follow!(other_user)
+          visit user_path(other_user)
+        end
 
-			describe "unfollowing a user" do
-				before do
-					user.follow!(other_user)
-					visit user_path(other_user)
-				end
+        it 'should decrement the followed user count' do
+          expect do
+            click_button 'Unfollow'
+          end.to change(user.followed_users, :count).by(-1)
+        end
 
-				it "should decrement the followed user count" do
-					expect do
-						click_button "Unfollow"
-					end.to change(user.followed_users, :count).by(-1)
-				end
+        it "should decrement the other user's followers count" do
+          expect do
+            click_button 'Unfollow'
+          end.to change(other_user.followers, :count).by(-1)
+        end
 
-				it "should decrement the other user's followers count" do
-					expect do
-						click_button "Unfollow"
-					end.to change(other_user.followers, :count).by(-1)
-				end
+        describe 'toggling the button' do
+          before { click_button 'Unfollow' }
+          it { should have_xpath("//input[@value='Follow']") }
+        end
+      end
+    end
+  end
 
-				describe "toggling the button" do
-					before { click_button "Unfollow" }
-					it { should have_xpath("//input[@value='Follow']") }
-				end
-			end
-		end
-	end
+  describe 'signup page' do
+    before { visit signup_path }
 
+    it { should have_content('Sign up') }
+    it { should have_title(full_title('Sign up')) }
+  end
 
-	describe "signup page" do
-		before { visit signup_path }
+  describe 'signup' do
+    before { visit signup_path }
 
-		it { should have_content('Sign up') }
-		it { should have_title(full_title('Sign up')) }
-	end
+    let(:submit) { 'Create my account' }
 
-	describe "signup" do
+    describe 'with invalid information' do
+      it 'should not create a user' do
+        expect { click_button submit }.not_to change(User, :count)
+      end
 
-		before { visit signup_path }
+      describe 'after submission' do
+        before { click_button submit }
 
-		let(:submit) { "Create my account" }
+        it { should have_title('Sign up') }
+        it { should have_content('error') }
+      end
+    end
 
-		describe "with invalid information" do
-			it "should not create a user" do
-				expect { click_button submit }.not_to change(User, :count)
-			end
+    describe 'with valid information' do
+      before do
+        fill_in 'Name', with: 'Example User'
+        fill_in 'Email',        with: 'user@example.com'
+        fill_in 'Password',     with: 'foobar'
+        fill_in 'Confirm Password', with: 'foobar'
+      end
 
-			describe "after submission" do
-				before { click_button submit }
+      it 'should create a user' do
+        expect { click_button submit }.to change(User, :count).by(1)
+      end
 
-				it { should have_title('Sign up') }
-				it { should have_content('error') }
-			end
-		end
+      describe 'after saving the user' do
+        before { click_button submit }
+        let(:user) { User.find_by(email: 'user@example.com') }
 
-		describe "with valid information" do
-			before do
-				fill_in "Name",         with: "Example User"
-				fill_in "Email",        with: "user@example.com"
-				fill_in "Password",     with: "foobar"
-				fill_in "Confirm Password", with: "foobar"
-			end
+        it { should have_title(user.name) }
+        it { should have_selector('div.alert.alert-success', text: 'Welcome') }
+      end
+    end
+  end
 
-			it "should create a user" do
-				expect { click_button submit }.to change(User, :count).by(1)
-			end
+  describe 'edit' do
+    let(:user) { FactoryBot.create(:user) }
+    before do
+      sign_in user
+      visit edit_user_path(user)
+    end
 
-			describe "after saving the user" do
-				before { click_button submit }
-				let(:user) { User.find_by(email: 'user@example.com') }
+    describe 'page' do
+      it { should have_content('Update your profile') }
+      it { should have_title('Edit user') }
+      it { should have_link('change', href: 'http://gravatar.com/emails') }
+    end
 
-				it { should have_title(user.name) }
-				it { should have_selector('div.alert.alert-success', text: 'Welcome') }
-			end
-		end
-	end
+    describe 'with invalid information' do
+      before { click_button 'Save changes' }
 
-	describe "edit" do
-		let(:user) { FactoryGirl.create(:user) }
-		before do
-			sign_in user
-			visit edit_user_path(user)
-		end
+      it { should have_content('error') }
+    end
 
-		describe "page" do
-			it { should have_content("Update your profile") }
-			it { should have_title("Edit user") }
-			it { should have_link('change', href: 'http://gravatar.com/emails') }
-		end
+    describe 'with valid information' do
+      let(:new_name) { 'New Name' }
+      let(:new_email) { 'new@example.com' }
+      before do
+        fill_in 'Name', with: new_name
+        fill_in 'Email',            with: new_email
+        fill_in 'Password',         with: user.password
+        fill_in 'Confirm Password', with: user.password
+        click_button 'Save changes'
+      end
 
-		describe "with invalid information" do
-			before { click_button "Save changes" }
+      it { should have_title(new_name) }
+      it { should have_selector('div.alert.alert-success') }
+      it { should have_link('Sign out', href: signout_path) }
+      # This reloads the user variable from the test database using user.reload,
+      # and then verifies that the user’s new name and email match the new values.
+      specify { expect(user.reload.name).to  eq new_name }
+      specify { expect(user.reload.email).to eq new_email }
+    end
 
-			it { should have_content('error') }
-		end
+    describe 'forbidden attributes', type: :request do
+      let(:params) do
+        { user: { admin: true, password: user.password,
+                  password_confirmation: user.password } }
+      end
+      before do
+        sign_in user, no_capybara: true
+        patch user_path(user), params: params
+      end
+      specify { expect(user.reload).not_to be_admin }
+    end
+  end
 
-		describe "with valid information" do
-			let(:new_name)  { "New Name" }
-			let(:new_email) { "new@example.com" }
-			before do
-				fill_in "Name",             with: new_name
-				fill_in "Email",            with: new_email
-				fill_in "Password",         with: user.password
-				fill_in "Confirm Password", with: user.password
-				click_button "Save changes"
-			end
+  describe 'index' do
+    let(:user) { FactoryBot.create(:user) }
+    # uses before(:each) to emphasize the contrast with before(:all).
+    before(:each) do
+      sign_in user
+      visit users_path
+    end
 
-			it { should have_title(new_name) }
-			it { should have_selector('div.alert.alert-success') }
-			it { should have_link('Sign out', href: signout_path) }
-			# This reloads the user variable from the test database using user.reload, 
-			# and then verifies that the user’s new name and email match the new values.
-			specify { expect(user.reload.name).to  eq new_name }
-			specify { expect(user.reload.email).to eq new_email }
-		end
+    it { should have_title('All users') }
+    it { should have_content('All users') }
 
+    # it "should list each user" do
+    # 	User.all.each do |user|
+    # 		expect(page).to have_selector('li', text: user.name)
+    # 	end
+    # end
 
-		describe "forbidden attributes" do
-			let(:params) do
-				{ user: { admin: true, password: user.password,
-					password_confirmation: user.password } }
-				end
-				before do
-					sign_in user, no_capybara: true
-					patch user_path(user), params
-				end
-				specify { expect(user.reload).not_to be_admin }
-			end
-		end
+    describe 'pagination' do
+      # Note here the use of before(:all), which ensures that the sample users are created once,
+      # before all the tests in the block. This is an optimization for speed,
+      # as creating 30 users can be slow on some systems.
+      # We use the complementary method after(:all) to delete the users once we’re done.
+      before(:all) { 30.times { FactoryBot.create(:user) } }
+      after(:all) { User.delete_all }
 
-		describe "index" do
+      it { should have_selector('div.pagination') }
 
+      it 'should list each user' do
+        User.paginate(page: 1).each do |user|
+          expect(page).to have_selector('li', text: user.name)
+        end
+      end
+    end
 
-			let(:user) { FactoryGirl.create(:user) }
-		# uses before(:each) to emphasize the contrast with before(:all).
-		before(:each) do
-			sign_in user
-			visit users_path
-		end
+    describe 'delete links' do
+      it { should_not have_link('delete') }
 
-		it { should have_title('All users') }
-		it { should have_content('All users') }
+      describe 'as an admin user' do
+        let(:admin) { FactoryBot.create(:admin) }
+        before do
+          sign_in admin
+          visit users_path
+        end
 
-		# it "should list each user" do
-		# 	User.all.each do |user|
-		# 		expect(page).to have_selector('li', text: user.name)
-		# 	end
-		# end
+        it { should have_link('delete', href: user_path(User.first)) }
+        it 'should be able to delete another user' do
+          expect do
+            click_link('delete', match: :first)
+          end.to change(User, :count).by(-1)
+        end
+        it { should_not have_link('delete', href: user_path(admin)) }
+      end
+    end
+  end
 
-		describe "pagination" do
-			# Note here the use of before(:all), which ensures that the sample users are created once, 
-			# before all the tests in the block. This is an optimization for speed, 
-			# as creating 30 users can be slow on some systems. 
-			# We use the complementary method after(:all) to delete the users once we’re done.
-			before(:all) { 30.times { FactoryGirl.create(:user) } }
-			after(:all)  { User.delete_all }
+  describe 'following/followers' do
+    let(:user) { FactoryBot.create(:user) }
+    let(:other_user) { FactoryBot.create(:user) }
+    before { user.follow!(other_user) }
 
-			it { should have_selector('div.pagination') }
+    describe 'followed users' do
+      before do
+        sign_in user
+        visit following_user_path(user)
+      end
 
-			it "should list each user" do
-				User.paginate(page: 1).each do |user|
-					expect(page).to have_selector('li', text: user.name)
-				end
-			end
-		end
+      it { should have_title(full_title('Following')) }
+      it { should have_selector('h3', text: 'Following') }
+      it { should have_link(other_user.name, href: user_path(other_user)) }
+    end
 
-		describe "delete links" do
+    describe 'followers' do
+      before do
+        sign_in other_user
+        visit followers_user_path(other_user)
+      end
 
-			it { should_not have_link('delete') }
-
-			describe "as an admin user" do
-				let(:admin) { FactoryGirl.create(:admin) }
-				before do
-					sign_in admin
-					visit users_path
-				end
-
-				it { should have_link('delete', href: user_path(User.first)) }
-				it "should be able to delete another user" do
-					expect do
-						click_link('delete', match: :first)
-					end.to change(User, :count).by(-1)
-				end
-				it { should_not have_link('delete', href: user_path(admin)) }
-			end
-		end
-	end
-
-	describe "following/followers" do
-		let(:user) { FactoryGirl.create(:user) }
-		let(:other_user) { FactoryGirl.create(:user) }
-		before { user.follow!(other_user) }
-
-		describe "followed users" do
-			before do
-				sign_in user
-				visit following_user_path(user)
-			end
-
-			it { should have_title(full_title('Following')) }
-			it { should have_selector('h3', text: 'Following') }
-			it { should have_link(other_user.name, href: user_path(other_user)) }
-		end
-
-		describe "followers" do
-			before do
-				sign_in other_user
-				visit followers_user_path(other_user)
-			end
-
-			it { should have_title(full_title('Followers')) }
-			it { should have_selector('h3', text: 'Followers') }
-			it { should have_link(user.name, href: user_path(user)) }
-		end
-	end
+      it { should have_title(full_title('Followers')) }
+      it { should have_selector('h3', text: 'Followers') }
+      it { should have_link(user.name, href: user_path(user)) }
+    end
+  end
 end
