@@ -1,5 +1,4 @@
 class StaticPagesController < ApplicationController
-  
   def home
     if signed_in?
       @micropost  = current_user.microposts.build
@@ -7,12 +6,9 @@ class StaticPagesController < ApplicationController
     end
   end
 
-  def help
-  end
+  def help; end
 
-  def about
-  end
+  def about; end
 
-  def contact
-  end
+  def contact; end
 end
