@@ -1,30 +1,30 @@
-Given /^a user visits the signin page$/ do
+Given(/^a user visits the signin page$/) do
   visit signin_path
 end
 
-When /^they submit invalid signin information$/ do
-  click_button "Sign in"
+When(/^they submit invalid signin information$/) do
+  click_button 'Sign in'
 end
 
-Then /^they should see an error message$/ do
+Then(/^they should see an error message$/) do
   expect(page).to have_selector('div.alert.alert-error')
 end
 
-Given /^the user has an account$/ do
-  @user = User.create(name: "Example User", email: "user@example.com",
-                      password: "foobar", password_confirmation: "foobar")
+Given(/^the user has an account$/) do
+  @user = User.create(name: 'Example User', email: 'user@example.com',
+                      password: 'foobar', password_confirmation: 'foobar')
 end
 
-When /^the user submits valid signin information$/ do
-  fill_in "Email",    with: @user.email
-  fill_in "Password", with: @user.password
-  click_button "Sign in"
+When(/^the user submits valid signin information$/) do
+  fill_in 'Email',    with: @user.email
+  fill_in 'Password', with: @user.password
+  click_button 'Sign in'
 end
 
-Then /^they should see their profile page$/ do
+Then(/^they should see their profile page$/) do
   expect(page).to have_title(@user.name)
 end
 
-Then /^they should see a signout link$/ do
+Then(/^they should see a signout link$/) do
   expect(page).to have_link('Sign out', href: signout_path)
 end
