@@ -1,215 +1,214 @@
 require 'spec_helper'
 
 describe User do
+  before do
+    @user = User.new(name: 'Example User', email: 'user@example.com', password: 'foobar',
+                     password_confirmation: 'foobar')
+  end
 
-	before do
-		@user = User.new(name: "Example User", email: "user@example.com", password: "foobar", password_confirmation: "foobar")
-	end
+  subject { @user }
 
-	subject { @user }
+  it { should respond_to(:name) }
+  it { should respond_to(:email) }
+  it { should respond_to(:password_digest) }
+  it { should respond_to(:password) }
+  it { should respond_to(:password_confirmation) }
+  it { should respond_to(:authenticate) }
+  it { should respond_to(:admin) }
+  it { should respond_to(:microposts) }
+  it { should respond_to(:feed) }
+  it { should respond_to(:relationships) }
+  it { should respond_to(:followed_users) }
+  # To create a following relationship, we’ll introduce a follow! utility method so that we can write user.follow!(other_user).
+  # (This follow! method should always work, so, as with create! and save!,
+  # we indicate with an exclamation point that an exception will be raised on failure.)
+  # We’ll also add an associated following? boolean method to test if one user is following another.
+  # The tests in Listing 11.11 show how we expect these methods to be used in practice.
+  it { should respond_to(:following?) }
+  it { should respond_to(:follow!) }
+  it { should respond_to(:reverse_relationships) }
+  it { should respond_to(:followers) }
 
-	it { should respond_to(:name) }
-	it { should respond_to(:email) }
-	it { should respond_to(:password_digest) }
-	it { should respond_to(:password) }
-	it { should respond_to(:password_confirmation) }
-	it { should respond_to(:authenticate) }
-	it { should respond_to(:admin) }
-	it { should respond_to(:microposts) }
-	it { should respond_to(:feed) }
-	it { should respond_to(:relationships) }
-	it { should respond_to(:followed_users) }
-	# To create a following relationship, we’ll introduce a follow! utility method so that we can write user.follow!(other_user). 
-	# (This follow! method should always work, so, as with create! and save!, 
-	# we indicate with an exclamation point that an exception will be raised on failure.) 
-	# We’ll also add an associated following? boolean method to test if one user is following another.
-	# The tests in Listing 11.11 show how we expect these methods to be used in practice.
-	it { should respond_to(:following?) }
-	it { should respond_to(:follow!) }
-	it { should respond_to(:reverse_relationships) }
-	it { should respond_to(:followers) }
+  it { should be_valid }
+  it { should_not be_admin }
 
-	it { should be_valid }
-	it { should_not be_admin }
+  describe 'when name is not present' do
+    before { @user.name = ' ' }
+    it { should_not be_valid }
+  end
 
-	describe "when name is not present" do
-		before { @user.name = " " }
-		it { should_not be_valid }
-	end
+  describe 'when email is not present' do
+    before { @user.email = ' ' }
+    it { should_not be_valid }
+  end
 
-	describe "when email is not present" do
-		before { @user.email = " " }
-		it { should_not be_valid }
-	end
+  describe 'when name is too long' do
+    before { @user.name = 'a' * 51 }
+    it { should_not be_valid }
+  end
 
-	describe "when name is too long" do
-		before { @user.name = "a" * 51 }
-		it { should_not be_valid }
-	end
+  describe 'when email format is invalid' do
+    it 'should be invalid' do
+      addresses = %w[user@foo,com user_at_foo.org example.user@foo. foo@bar_baz.com foo@bar+baz.com]
+      addresses.each do |invalid_address|
+        @user.email = invalid_address
+        expect(@user).not_to be_valid
+      end
+    end
+  end
 
-	describe "when email format is invalid" do
-		it "should be invalid" do
-			addresses = %w[user@foo,com user_at_foo.org example.user@foo. foo@bar_baz.com foo@bar+baz.com]
-			addresses.each do |invalid_address|
-				@user.email = invalid_address
-				expect(@user).not_to be_valid
-			end
-		end
-	end
+  describe 'when email format is valid' do
+    it 'should be valid' do
+      addresses = %w[user@foo.COM A_US-ER@f.b.org frst.lst@foo.jp a+b@baz.cn]
+      addresses.each do |valid_address|
+        @user.email = valid_address
+        expect(@user).to be_valid
+      end
+    end
+  end
 
-	describe "when email format is valid" do
-		it "should be valid" do
-			addresses = %w[user@foo.COM A_US-ER@f.b.org frst.lst@foo.jp a+b@baz.cn]
-			addresses.each do |valid_address|
-				@user.email = valid_address
-				expect(@user).to be_valid
-			end
-		end
-	end	
+  describe 'when email address is already taken' do
+    before do
+      user_with_same_email = @user.dup
+      user_with_same_email.email = @user.email.upcase
+      user_with_same_email.save
+    end
 
-	describe "when email address is already taken" do
-		before do
-			user_with_same_email = @user.dup
-			user_with_same_email.email = @user.email.upcase
-			user_with_same_email.save
-		end
+    it { should_not be_valid }
+  end
 
-		it { should_not be_valid }
-	end
+  describe 'email address with mixed case' do
+    let(:mixed_case_email) { 'Foo@ExAMPle.CoM' }
 
-	describe "email address with mixed case" do
-		let(:mixed_case_email) { "Foo@ExAMPle.CoM" }
+    it 'should be saved as all lower-case' do
+      @user.email = mixed_case_email
+      @user.save
+      expect(@user.reload.email).to eq mixed_case_email.downcase
+    end
+  end
 
-		it "should be saved as all lower-case" do
-			@user.email = mixed_case_email
-			@user.save
-			expect(@user.reload.email).to eq mixed_case_email.downcase
-		end
-	end
-	
-	# Password
-	describe "when password is not present" do
-		before do
-			@user = User.new(name: "Example User", email: "user@example.com", password: " ", password_confirmation: " ")
-		end
-		it { should_not be_valid }
-	end
+  # Password
+  describe 'when password is not present' do
+    before do
+      @user = User.new(name: 'Example User', email: 'user@example.com', password: ' ', password_confirmation: ' ')
+    end
+    it { should_not be_valid }
+  end
 
-	describe "when password doesn't match confirmation" do
-		before { @user.password_confirmation = "mismatch" }
-		it { should_not be_valid }
-	end
+  describe "when password doesn't match confirmation" do
+    before { @user.password_confirmation = 'mismatch' }
+    it { should_not be_valid }
+  end
 
-	describe "with a password that's too short" do
-		before { @user.password = @user.password_confirmation = "a" * 5 }
-		it { should be_invalid }
-	end
+  describe "with a password that's too short" do
+    before { @user.password = @user.password_confirmation = 'a' * 5 }
+    it { should be_invalid }
+  end
 
-	describe "return value of authenticate method" do
-		before { @user.save }
-		let(:found_user) { User.find_by(email: @user.email) }
+  describe 'return value of authenticate method' do
+    before { @user.save }
+    let(:found_user) { User.find_by(email: @user.email) }
 
-		describe "with valid password" do
-			it { should eq found_user.authenticate(@user.password) }
-		end
+    describe 'with valid password' do
+      it { should eq found_user.authenticate(@user.password) }
+    end
 
-		describe "with invalid password" do
-			let(:user_for_invalid_password) { found_user.authenticate("invalid") }
+    describe 'with invalid password' do
+      let(:user_for_invalid_password) { found_user.authenticate('invalid') }
 
-			it { should_not eq user_for_invalid_password }
-			specify { expect(user_for_invalid_password).to be_false }
-		end
-	end
+      it { should_not eq user_for_invalid_password }
+      specify { expect(user_for_invalid_password).to be_false }
+    end
+  end
 
+  describe 'remember token' do
+    before { @user.save }
+    its(:remember_token) { should_not be_blank }
+  end
 
-	describe "remember token" do
-		before { @user.save }
-		its(:remember_token) { should_not be_blank }
-	end
+  describe "with admin attribute set to 'true'" do
+    before do
+      @user.save!
+      @user.toggle!(:admin)
+    end
 
-	describe "with admin attribute set to 'true'" do
-		before do
-			@user.save!
-			@user.toggle!(:admin)
-		end
+    it { should be_admin }
+  end
 
-		it { should be_admin }
-	end
+  describe 'micropost associations' do
+    #  the reason is that let variables are lazy, meaning that
+    # they only spring into existence when referenced.
+    # The problem is that we want the microposts to exist immediately,
+    # so that the timestamps are in the right order and so that
+    # @user.microposts isn’t empty. We accomplish this with let!,
+    # which forces the corresponding variable to come into existence immediately.
 
-	describe "micropost associations" do
-		#  the reason is that let variables are lazy, meaning that 
-		# they only spring into existence when referenced. 
-		# The problem is that we want the microposts to exist immediately, 
-		# so that the timestamps are in the right order and so that
-		# @user.microposts isn’t empty. We accomplish this with let!, 
-		# which forces the corresponding variable to come into existence immediately.
+    before { @user.save }
+    let!(:older_micropost) do
+      FactoryBot.create(:micropost, user: @user, created_at: 1.day.ago)
+    end
+    let!(:newer_micropost) do
+      FactoryBot.create(:micropost, user: @user, created_at: 1.hour.ago)
+    end
 
-		before { @user.save }
-		let!(:older_micropost) do
-			FactoryGirl.create(:micropost, user: @user, created_at: 1.day.ago)
-		end
-		let!(:newer_micropost) do
-			FactoryGirl.create(:micropost, user: @user, created_at: 1.hour.ago)
-		end
+    it 'should have the right microposts in the right order' do
+      expect(@user.microposts.to_a).to eq [newer_micropost, older_micropost]
+    end
 
-		it "should have the right microposts in the right order" do
-			expect(@user.microposts.to_a).to eq [newer_micropost, older_micropost]
-		end
+    it 'should destroy associated microposts' do
+      microposts = @user.microposts.to_a
+      @user.destroy
+      expect(microposts).not_to be_empty
+      # Here we have used Micropost.where instead of Micropost.
+      # find because it returns an empty object if the record is not found instead of raising an exception,
+      # which is a little easier to test.
+      microposts.each do |micropost|
+        expect(Micropost.where(id: micropost.id)).to be_empty
+      end
+    end
 
-		it "should destroy associated microposts" do
-			microposts = @user.microposts.to_a
-			@user.destroy
-			expect(microposts).not_to be_empty
-			# Here we have used Micropost.where instead of Micropost.
-			# find because it returns an empty object if the record is not found instead of raising an exception, 
-			# which is a little easier to test.
-			microposts.each do |micropost|
-				expect(Micropost.where(id: micropost.id)).to be_empty
-			end
-		end
+    describe 'status' do
+      let(:unfollowed_post) do
+        FactoryBot.create(:micropost, user: FactoryBot.create(:user))
+      end
+      let(:followed_user) { FactoryBot.create(:user) }
 
-		describe "status" do
-			let(:unfollowed_post) do
-				FactoryGirl.create(:micropost, user: FactoryGirl.create(:user))
-			end
-			let(:followed_user) { FactoryGirl.create(:user) }
+      before do
+        @user.follow!(followed_user)
+        3.times { followed_user.microposts.create!(content: 'Lorem ipsum') }
+      end
 
-			before do
-				@user.follow!(followed_user)
-				3.times { followed_user.microposts.create!(content: "Lorem ipsum") }
-			end
+      its(:feed) { should include(newer_micropost) }
+      its(:feed) { should include(older_micropost) }
+      its(:feed) { should_not include(unfollowed_post) }
+      its(:feed) do
+        followed_user.microposts.each do |micropost|
+          should include(micropost)
+        end
+      end
+    end
+  end
 
-			its(:feed) { should include(newer_micropost) }
-			its(:feed) { should include(older_micropost) }
-			its(:feed) { should_not include(unfollowed_post) }
-			its(:feed) do
-				followed_user.microposts.each do |micropost|
-					should include(micropost)
-				end
-			end
-		end
-	end
+  describe 'following' do
+    let(:other_user) { FactoryBot.create(:user) }
+    before do
+      @user.save
+      @user.follow!(other_user)
+    end
 
-	describe "following" do
-		let(:other_user) { FactoryGirl.create(:user) }
-		before do
-			@user.save
-			@user.follow!(other_user)
-		end
+    it { should be_following(other_user) }
+    its(:followed_users) { should include(other_user) }
 
-		it { should be_following(other_user) }
-		its(:followed_users) { should include(other_user) }
+    describe 'and unfollowing' do
+      before { @user.unfollow!(other_user) }
 
-		describe "and unfollowing" do
-			before { @user.unfollow!(other_user) }
+      it { should_not be_following(other_user) }
+      its(:followed_users) { should_not include(other_user) }
+    end
 
-			it { should_not be_following(other_user) }
-			its(:followed_users) { should_not include(other_user) }
-		end
-
-		describe "followed user" do
-			subject { other_user }
-			its(:followers) { should include(@user) }
-		end
-	end
+    describe 'followed user' do
+      subject { other_user }
+      its(:followers) { should include(@user) }
+    end
+  end
 end
