@@ -1,9 +1,9 @@
-include ApplicationHelper
+# include ApplicationHelper
 
 def valid_signin(user)
-  fill_in "Email",    with: user.email
-  fill_in "Password", with: user.password
-  click_button "Sign in"
+  fill_in 'Email',    with: user.email
+  fill_in 'Password', with: user.password
+  click_button 'Sign in'
 end
 
 RSpec::Matchers.define :have_error_message do |message|
@@ -13,7 +13,7 @@ RSpec::Matchers.define :have_error_message do |message|
 end
 
 def full_title(page_title)
-  base_title = "Ruby on Rails Tutorial Sample App"
+  base_title = 'Ruby on Rails Tutorial Sample App'
   if page_title.empty?
     base_title
   else
@@ -21,8 +21,7 @@ def full_title(page_title)
   end
 end
 
-
-def sign_in(user, options={})
+def sign_in(user, options = {})
   if options[:no_capybara]
     # Sign in when not using Capybara.
     remember_token = User.new_remember_token
@@ -30,9 +29,8 @@ def sign_in(user, options={})
     user.update_attribute(:remember_token, User.encrypt(remember_token))
   else
     visit signin_path
-    fill_in "Email",    with: user.email
-    fill_in "Password", with: user.password
-    click_button "Sign in"
+    fill_in 'Email',    with: user.email
+    fill_in 'Password', with: user.password
+    click_button 'Sign in'
   end
 end
-
